backend/pyproject.toml
[project]
name = "roomly-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.31.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.128.8",
    "fastapi-users[sqlalchemy]>=14.0.2",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.33",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
